refactor: rename settings file
