+++ app/utils/settings.js
